modify: Service 인터페이스 삭제
 - 사실상 쓸 이유가 없는 Service 인터페이스들을 삭제
 - 각 Impl 클래스들의 이름을 Service로 변경
