+++ src/main/java/subway/line/LineService.java
@@ -1,19 +1,47 @@
 package subway.line;
 
+import org.springframework.stereotype.Service;
 import subway.section.Section;
+import subway.section.SectionDao;
 
 import java.util.List;
 
-public interface LineService {
-    Line save(Line line, Section section);
+@Service
+public class LineService{
+    private final LineDao lineDao;
+    private final SectionDao sectionDao;
 
-    boolean deleteById(Long lineId);
+    public LineService(LineDao lineDao, SectionDao sectionDao) {
+        this.lineDao = lineDao;
+        this.sectionDao = sectionDao;
+    }
 
-    List<Line> findAll();
+    public Line save(Line line, Section section) {
+        Line newLine = lineDao.save(line);
+        if (newLine != null) {
+            sectionDao.save(new Section(section.getUpStationId(), section.getDownStationId(), section.getDistance(), newLine.getId()));
+        }
+        return newLine;
+    }
 
-    Line findOne(Long lineId);
+    public boolean deleteById(Long lineId) {
+        return lineDao.deleteById(lineId) != 0;
+    }
 
-    boolean update(Line line);
+    public List<Line> findAll() {
+        return lineDao.findAll();
+    }
 
-    boolean updateAll(Line line);
+    public Line findOne(Long lineId) {
+        return lineDao.findOne(lineId);
+    }
+
+    public boolean update(Line line) {
+        return lineDao.update(line) != 0;
+    }
+
+    public boolean updateAll(Line line) {
+        return lineDao.updateAll(line) != 0;
+    }
+
 }
