+++ src/main/java/subway/section/SectionController.java
@@ -3,7 +3,6 @@ package subway.section;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import subway.line.LineService;
 
 @RestController
 @RequestMapping("/lines")
